requirements.txt
openai>=1.13.3
pyyaml
cx_Freeze
ttkbootstrap==0.5.1
pillow
playwright
chardet
requests